refactor - Distance, Distances 잘못된 위치 수정
